dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi-cli>=0.0.16",
    "fastapi-users[sqlalchemy]>=14.0.2",
    "fastapi[standard]>=0.123.5",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
